+++ src/main/java/com/bina/varsim/tools/evaluation/VCFcompare.java
@@ -5,6 +5,9 @@ import com.bina.varsim.intervalTree.SimpleInterval1D;
 import com.bina.varsim.intervalTree.ValueInterval1D;
 import com.bina.varsim.types.*;
 import com.bina.varsim.types.stats.EnumStatsRatioCounter;
+import com.bina.varsim.types.variant.Variant;
+import com.bina.varsim.types.variant.VariantOverallType;
+import com.bina.varsim.types.variant.VariantType;
 import com.bina.varsim.util.VCFparser;
 import com.bina.varsim.util.chrSearchTree;
 import com.fasterxml.jackson.core.JsonGenerator;
