+++ src/main/java/com/bina/varsim/tools/simulation/RandDGV2VCF.java
@@ -2,6 +2,7 @@ package com.bina.varsim.tools.simulation;
 
 import com.bina.varsim.constants.Constant;
 import com.bina.varsim.types.*;
+import com.bina.varsim.types.variant.Variant;
 import com.bina.varsim.util.DGVparser;
 import com.bina.varsim.util.SimpleReference;
 import org.apache.log4j.Logger;
