--- src/main/java/com/bina/varsim/types/Constraint.java
@@ -1,8 +1,0 @@
-package com.bina.varsim.types;
-
-/**
- * Created by johnmu on 5/20/15.
- */
-public class Constraint {
-
-}
