+++ src/main/java/com/bina/varsim/types/constraint/Constraint.java
@@ -1,0 +1,10 @@
+package com.bina.varsim.types.constraint;
+
+/**
+ * Created by johnmu on 5/20/15.
+ */
+public class Constraint {
+    public Constraint(String constraintArg) {
+        
+    }
+}
