+++ src/main/java/com/bina/varsim/util/VCFparser.java
@@ -4,7 +4,7 @@ package com.bina.varsim.util;
 
 import com.bina.varsim.types.ChrString;
 import com.bina.varsim.types.FlexSeq;
-import com.bina.varsim.types.Variant;
+import com.bina.varsim.types.variant.Variant;
 import org.apache.log4j.Logger;
 
 import java.io.BufferedReader;
